// eshop-main/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type User = {
  __typename?: 'User';
  id: string;
  email: string;
  role: string;
  firstName?: Maybe<string>;
  lastName?: Maybe<string>;
};

export type AuthPayload = {
  __typename?: 'AuthPayload';
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
  user: User;
};

export type Scalars = {
  ID: string;
  String: string;
  Int: number;
  Float: number;
  Boolean: boolean;
  JSON: Record<string, unknown>;
};

export type GetProductsQuery = {
  products?: Maybe<ProductConnection>;
};

export type GetProductsQueryVariables = Exact<{
  first?: InputMaybe<number>;
  after?: InputMaybe<string>;
}>;

export type GetProductQuery = {
  product?: Maybe<Product>;
};

export type GetProductQueryVariables = Exact<{ slug: Scalars['String'] }>;

export type GetMeQuery = { me?: Maybe<User> };

export type GetMeQueryVariables = Exact<{ [key: string]: never }>;

export type GetCartQuery = { cart?: Maybe<Cart> };

export type GetCartQueryVariables = Exact<{ [key: string]: never }>;

export type GetOrdersQuery = { orders?: Maybe<OrderConnection> };

export type GetOrdersQueryVariables = Exact<{ [key: string]: never }>;

export type LoginMutation = { login: AuthPayload };

export type LoginMutationVariables = Exact<{
  email: Scalars['String'];
  password: Scalars['String'];
}>;

export type RegisterMutation = { register: AuthPayload };

export type RegisterMutationVariables = Exact<{
  email: Scalars['String'];
  password: Scalars['String'];
  firstName?: InputMaybe<Scalars['String']>;
  lastName?: InputMaybe<Scalars['String']>;
}>;
// ...
export const GetMeDocument = gql`
  query GetMe {
    me {
      id
      email
      role
      firstName
      lastName
    }
  }
` as unknown as TypedDocumentNode<GetMeQuery, GetMeQueryVariables>;
// ...
export const LoginDocument = gql`
  mutation Login($email: String!, $password: String!) {
    login(input: { email: $email, password: $password }) {
      accessToken
      refreshToken
      expiresIn
      user {
        id
        email
        role
      }
    }
  }
` as unknown as TypedDocumentNode<LoginMutation, LoginMutationVariables>;

export const RegisterDocument = gql`
  mutation Register(
    $email: String!
    $password: String!
    $firstName: String
    $lastName: String
  ) {
    register(
      input: {
        email: $email
        password: $password
        firstName: $firstName
        lastName: $lastName
      }
    ) {
      accessToken
      refreshToken
      expiresIn
      user {
        id
        email
        role
      }
    }
  }
` as unknown as TypedDocumentNode<RegisterMutation, RegisterMutationVariables>;